requirements.txt
pillow
tk
qrcode